fix: fix issue with moved cleanUpListsWithAnnotations, move both clean up functions
