+++ packages/super-editor/src/extensions/field-annotation/cleanup-commands/index.js
@@ -1,0 +1,10 @@
+import * as cleanUpListsCommands from './cleanUpListsWithAnnotations.js';
+import * as cleanUpParagraphCommands from './cleanUpParagraphWithAnnotations.js';
+
+export const commands = {
+  ...cleanUpListsCommands,
+  ...cleanUpParagraphCommands,
+};
+
+export * from './cleanUpListsWithAnnotations.js';
+export * from './cleanUpParagraphWithAnnotations.js';
